fix: Optimize Dashboard Data Fetching (Student Dashboard)
