+++ web/components/dashboard/student-dashboard-content.tsx
@@ -21,9 +21,8 @@ import {
 import { Button } from "@/components/ui/button";
 import { Brain } from "lucide-react";
 import { useRouter } from "next/navigation";
-import { VelocityMetrics } from "@/server/services/metrics/velocity-service";
-import { GenreMetricsResponse } from "@/server/services/metrics/genre-engagement-service";
 import { useDashboardTelemetry } from "@/lib/telemetry/dashboard-telemetry";
+import { useDashboardMetrice } from "@/hooks/student/useDashboardMetrice"
 
 interface StudentDashboardContentProps {
   userId: string;
@@ -37,14 +36,6 @@ interface StudentDashboardContentProps {
   };
 }
 
-interface DashboardData {
-  velocity: VelocityMetrics | null;
-  genres: GenreMetricsResponse | null;
-  srsHealth: any | null;
-  aiInsights: any | null;
-  activityTimeline: any | null;
-}
-
 export default function StudentDashboardContent({
   userId,
   user,
@@ -52,14 +43,7 @@ export default function StudentDashboardContent({
   const t = useScopedI18n("pages.student.dashboard") as any;
   const router = useRouter();
   const { trackEvent } = useDashboardTelemetry();
-  const [loading, setLoading] = React.useState(true);
-  const [data, setData] = React.useState<DashboardData>({
-    velocity: null,
-    genres: null,
-    srsHealth: null,
-    aiInsights: null,
-    activityTimeline: null,
-  });
+  const { data, loading, refresh } = useDashboardMetrice(userId);
 
   // Track dashboard view
   React.useEffect(() => {
@@ -69,86 +53,6 @@ export default function StudentDashboardContent({
       xp: user.xp,
     });
   }, [trackEvent, user]);
-
-  // Fetch all dashboard data
-  const fetchDashboardData = React.useCallback(async () => {
-    try {
-      setLoading(true);
-
-      // Fetch all metrics in parallel
-      const [velocityRes, genresRes, srsRes, aiRes, activityRes] =
-        await Promise.allSettled([
-          fetch(`/api/v1/metrics/velocity?studentId=${userId}`),
-          fetch(
-            `/api/v1/metrics/genres?studentId=${userId}&timeframe=30d&enhanced=true&includeRecommendations=true`
-          ),
-          fetch(`/api/v1/metrics/srs?studentId=${userId}&includeDetails=true`),
-          fetch(`/api/v1/ai/summary?userId=${userId}&kind=student`),
-          fetch(
-            `/api/v1/metrics/activity?entityId=${userId}&scope=student&format=timeline&timeframe=30d`
-          ),
-        ]);
-
-      // Process velocity data
-      if (velocityRes.status === "fulfilled" && velocityRes.value.ok) {
-        const velocityData = await velocityRes.value.json();
-
-        // Check if response has student scope structure
-        if (velocityData.scope === "student" && velocityData.student) {
-          setData((prev) => ({ ...prev, velocity: velocityData.student }));
-        } else {
-          console.error("Unexpected velocity data structure:", velocityData);
-        }
-      } else if (velocityRes.status === "fulfilled") {
-        console.error("Velocity fetch failed:", await velocityRes.value.text());
-      }
-
-      // Process genre data
-      if (genresRes.status === "fulfilled" && genresRes.value.ok) {
-        const genreData = await genresRes.value.json();
-        setData((prev) => ({ ...prev, genres: genreData }));
-      } else if (genresRes.status === "fulfilled") {
-        console.error("Genre fetch failed:", await genresRes.value.text());
-      }
-
-      // Process SRS health data
-      if (srsRes.status === "fulfilled" && srsRes.value.ok) {
-        const srsData = await srsRes.value.json();
-
-        // Check if response has student scope structure
-        if (srsData.scope === "student" && srsData.student) {
-          setData((prev) => ({ ...prev, srsHealth: srsData }));
-        } else {
-          console.error(
-            "Unexpected SRS data structure - expected student scope:",
-            srsData
-          );
-        }
-      } else if (srsRes.status === "fulfilled") {
-        console.error("SRS fetch failed:", await srsRes.value.text());
-      }
-
-      // Process AI insights (optional)
-      if (aiRes.status === "fulfilled" && aiRes.value.ok) {
-        const aiData = await aiRes.value.json();
-        setData((prev) => ({ ...prev, aiInsights: aiData }));
-      }
-
-      // Process activity timeline
-      if (activityRes.status === "fulfilled" && activityRes.value.ok) {
-        const activityData = await activityRes.value.json();
-        setData((prev) => ({ ...prev, activityTimeline: activityData }));
-      }
-    } catch (error) {
-      console.error("Error fetching dashboard data:", error);
-    } finally {
-      setLoading(false);
-    }
-  }, [userId]);
-
-  React.useEffect(() => {
-    fetchDashboardData();
-  }, [fetchDashboardData]);
 
   const handlePracticeFlashcards = () => {
     router.push("/student/vocabulary");
@@ -170,21 +74,21 @@ export default function StudentDashboardContent({
         <XPVelocityWidget
           data={data.velocity}
           loading={loading}
-          onRefresh={fetchDashboardData}
+          onRefresh={refresh}
         />
 
         {/* ETA Card - conditionally rendered */}
         <ETACard
           data={data.velocity}
           loading={loading}
-          onRefresh={fetchDashboardData}
+          onRefresh={refresh}
         />
 
         {/* Genre Engagement */}
         <GenreEngagementWidget
           data={data.genres}
           loading={loading}
-          onRefresh={fetchDashboardData}
+          onRefresh={refresh}
           onGenreClick={handleGenreClick}
         />
 
@@ -210,44 +114,44 @@ export default function StudentDashboardContent({
           data={
             data.srsHealth?.student
               ? {
-                  userId: data.srsHealth.student.userId,
-                  healthStatus:
-                    data.srsHealth.student.isOverloaded ||
+                userId: data.srsHealth.student.userId,
+                healthStatus:
+                  data.srsHealth.student.isOverloaded ||
                     data.srsHealth.student.hasCriticalBacklog
-                      ? "critical"
-                      : data.srsHealth.student.totalDueForReview > 20
-                        ? "moderate"
-                        : "healthy",
-                  metrics: {
-                    totalCards: data.srsHealth.student.totalCards,
-                    dueToday: data.srsHealth.student.totalDueForReview,
-                    overdue: data.srsHealth.student.totalOverdue,
-                    newCardsToday:
-                      data.srsHealth.student.vocabulary.new +
-                      data.srsHealth.student.sentences.new,
-                    reviewedToday: 0,
-                    avgRetentionRate:
-                      data.srsHealth.student.overallMasteryPct / 100,
-                  },
-                  recommendations:
-                    data.srsHealth.student.suggestedActions
-                      ?.slice(0, 3)
-                      .map((action: any) => ({
-                        action: action.title,
-                        priority: action.priority,
-                        reason: action.description,
-                      })) || [],
-                  quickActions:
-                    data.srsHealth.quickActions?.map((action: any) => ({
-                      label: action.title,
-                      count: action.targetCount || 0,
-                      action: action.type,
+                    ? "critical"
+                    : data.srsHealth.student.totalDueForReview > 20
+                      ? "moderate"
+                      : "healthy",
+                metrics: {
+                  totalCards: data.srsHealth.student.totalCards,
+                  dueToday: data.srsHealth.student.totalDueForReview,
+                  overdue: data.srsHealth.student.totalOverdue,
+                  newCardsToday:
+                    data.srsHealth.student.vocabulary.new +
+                    data.srsHealth.student.sentences.new,
+                  reviewedToday: 0,
+                  avgRetentionRate:
+                    data.srsHealth.student.overallMasteryPct / 100,
+                },
+                recommendations:
+                  data.srsHealth.student.suggestedActions
+                    ?.slice(0, 3)
+                    .map((action: any) => ({
+                      action: action.title,
+                      priority: action.priority,
+                      reason: action.description,
                     })) || [],
-                }
+                quickActions:
+                  data.srsHealth.quickActions?.map((action: any) => ({
+                    label: action.title,
+                    count: action.targetCount || 0,
+                    action: action.type,
+                  })) || [],
+              }
               : null
           }
           loading={loading}
-          onRefresh={fetchDashboardData}
+          onRefresh={refresh}
           onPracticeClick={handlePracticeFlashcards}
         />
 
@@ -265,15 +169,15 @@ export default function StudentDashboardContent({
           metrics={
             data.velocity
               ? {
-                  currentXp: data.velocity.currentXp,
-                  velocity: data.velocity.emaVelocity,
-                  genresRead: data.genres?.topGenres?.length || 0,
-                  retentionRate: data.srsHealth?.metrics?.avgRetentionRate || 0,
-                }
+                currentXp: data.velocity.currentXp,
+                velocity: data.velocity.emaVelocity,
+                genresRead: data.genres?.topGenres?.length || 0,
+                retentionRate: data.srsHealth?.metrics?.avgRetentionRate || 0,
+              }
               : undefined
           }
           loading={loading}
-          onRefresh={fetchDashboardData}
+          onRefresh={refresh}
         />
       </div>
     </div>
